📦 Separate Map and layers re-render
