+++ src/components/MapView/index.test.tsx
@@ -1,15 +1,10 @@
 import React from 'react';
-import { Provider } from 'react-redux';
 import { render } from '@testing-library/react';
-
 import MapView from '.';
-import { store } from '../../context/store';
+
+jest.mock('./Layers', () => 'mock-Layers');
 
 test('renders as expected', () => {
-  const { container } = render(
-    <Provider store={store}>
-      <MapView />
-    </Provider>,
-  );
+  const { container } = render(<MapView />);
   expect(container).toMatchSnapshot();
 });
